split out genotype loading into genotype module
