+++ src/Poseidon/CLI/Merge.hs
@@ -1,35 +1,36 @@
 module Poseidon.CLI.Merge (runMerge, MergeOptions(..)) where
 
-import           Poseidon.Package           (PoseidonPackage(..),
-                                            loadPoseidonPackages,
-                                            GenotypeDataSpec(..),
-                                            GenotypeFormatSpec(..),
-                                            ContributorSpec(..),
-                                            getJointGenotypeData, maybeLoadJannoFiles, maybeLoadBibTeXFiles) 
-import           Poseidon.Janno             (jannoToSimpleMaybeList,
-                                            writeJannoFile)
 import           Poseidon.BibFile           (bibToSimpleMaybeList,
-                                            writeBibTeXFile)
+                                             writeBibTeXFile)
+import           Poseidon.GenotypeData      (GenotypeDataSpec (..), GenotypeFormatSpec (..))
+import           Poseidon.Janno             (jannoToSimpleMaybeList,
+                                             writeJannoFile)
+import           Poseidon.Package           (ContributorSpec (..),
+                                             PoseidonPackage (..),
+                                             getJointGenotypeData,
+                                             loadPoseidonPackages,
+                                             maybeLoadBibTeXFiles,
+                                             maybeLoadJannoFiles)
 
 import           Control.Monad              (when)
 import           Data.Aeson                 (encodeFile)
 import           Data.List                  (nub, sortOn)
 import           Data.Maybe                 (catMaybes, isJust)
-import           Data.Time                  (UTCTime(..), getCurrentTime)
+import           Data.Time                  (UTCTime (..), getCurrentTime)
 import           Data.Version               (makeVersion)
 import           Pipes                      (runEffect, (>->))
 import           Pipes.Safe                 (runSafeT)
 import           SequenceFormats.Eigenstrat (writeEigenstrat)
-import           System.IO                  (hPutStrLn, stderr)
-import           System.FilePath            ((</>), (<.>))
 import           System.Directory           (createDirectory)
+import           System.FilePath            ((<.>), (</>))
+import           System.IO                  (hPutStrLn, stderr)
 import           Text.CSL.Reference         (refId)
 
 -- | A datatype representing command line options for the survey command
 data MergeOptions = MergeOptions
-    { _jaBaseDirs  :: [FilePath]
-    , _outPacPath  :: FilePath
-    , _outPacName  :: String
+    { _jaBaseDirs :: [FilePath]
+    , _outPacPath :: FilePath
+    , _outPacName :: String
     }
 
 -- | The main function running the janno command
